ref: Merge av-moderation logic back into ChatRoomImpl.
